complete refactor of code to be more coherent, clean up commenting, syntax. etc
